requirements.txt
# FastAPI framework & ASGI server
fastapi[all]        # Includes Starlette, Uvicorn, Pydantic, and essential dependencies
uvicorn[standard]   # ASGI server with logging & autoreload
slowapi             # Middleware rate limiter

# Database-related dependencies
asyncpg             # PostgreSQL async driver
sqlalchemy          # ORM for handling DB queries
alembic             # DB migrations
psycopg             # PostgreSQL driver

# Caching dependencies
redis               # Redis driver for caching
aioredis            # Async Redis client

# Other utilities
python-dotenv       # Load environment variables from .env
loguru              # Better logging
requests            # HTTP requests
httpx               # Async HTTP client

# Docker SDK (for container interactions if needed)
docker              # Python Docker SDK

# Testing dependencies 
pytest              # Testing framework
pytest-asyncio      # Async support for pytest
httpx               # API request testing
